Move services module to service_layer
